Make sim_cycle do the right thing.
Sequence is check inputs -> comb update -> seq update -> comb update.
Formerly the last comb update was excluded (the overall simulation was
correct, but the outputs per step would not always read back correctly
without the final comb update step).

The new functions allow the old behaviour to be written if needed (it is
a bit faster having only one comb update step).

A better solution would entail a bit of a rewrite to the scheduler to
properly calculate pre-comb and post-comb update steps.