Fixes to previous simulator update.
It worked in testbenches under the following assumptions;

* set up inputs
* run a cycle
* now we are after the cycle and can read new register outputs,
  including any combinatorial updates they caused.

The tracing code (vcd, waveterm etc) needs access to the output
values before the updates occured.  To do this the simulator
structure is now;

check_inputs -> comb_update_0 -> seq_update -> comb_update_1

This is what Cyclesim.Api.cycle now does (there is no longer
a specific cycle function in the structure).  When we want to
wrap the simulator for some reason (combining, checking or tracing)
we need to wrap one of the above functions, rather then the
whole cycle function as was done before, which gives us more
precise control over when it will occur.

??? Is the ability to read the new register value after 'cycle'
**really** correct or will it have some new problems?  I note
other levellized simulations split the cycle according to rising
and falling edges perhaps because both the outputs before the
register update and after it are relevant.  We can do this now,
but specific functions are not yet provided.  Possibly the following
apis are actually relevant;

- Sim.cycle1 (rising edge)
- Sim.cycle0 (falling edge)
- Sim.cycle ~n (n * (rise + fall))

cycle1 and cycle0 would implement everything needed. 'cycle', when
used, would be about 2x as efficient (the input read and last comb
update would not be needed).