add `matches` and `resize_*` functions
mainly for use in ppx_hardcaml, but also useful otherwise.

`matches` is similar to the `cases` function, but it more
focussed on the "sparse" case as it builds a tree structure
out of mux2's.

`resize_list` resizes a list of elements to have equal width.
`resize_op2` resizes 2 arguments and applies them to an operator.